+++ java/org/lwjgl/opencl/KHRGLSharing.java
@@ -176,7 +176,7 @@ public class KHRGLSharing {
 			checkNT(properties);
 			checkBufferSafe(param_value_size_ret, 1);
 		}
-		return nclGetGLContextInfoKHR(memAddress(properties), param_name, (long)remainingSafe(param_value), memAddressSafe(param_value), memAddressSafe(param_value_size_ret));
+		return nclGetGLContextInfoKHR(memAddress(properties), param_name, remainingSafe(param_value), memAddressSafe(param_value), memAddressSafe(param_value_size_ret));
 	}
 
 	/** PointerBuffer version of: {@link #clGetGLContextInfoKHR GetGLContextInfoKHR} */
